Updated module augmentation for ui package.
