+++ packages/ckeditor5-ui/src/augmentation.ts
@@ -3,11 +3,11 @@
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
+import type BalloonToolbar from './toolbar/balloon/balloontoolbar';
+import type BlockToolbar from './toolbar/block/blocktoolbar';
+import type ContextualBalloon from './panel/balloon/contextualballoon';
+import type Notification from './notification/notification';
 import type { ToolbarConfig } from '@ckeditor/ckeditor5-core';
-
-/**
- * @module ui/uiconfig
- */
 
 declare module '@ckeditor/ckeditor5-core' {
 	interface EditorConfig {
@@ -83,5 +83,12 @@ declare module '@ckeditor/ckeditor5-core' {
 		 * Read more about configuring the main editor toolbar in {@link module:core/editor/editorconfig~EditorConfig#toolbar}.
 		 */
 		blockToolbar?: ToolbarConfig;
+	}
+
+	interface PluginsMap {
+		[ BalloonToolbar.pluginName ]: BalloonToolbar;
+		[ BlockToolbar.pluginName ]: BlockToolbar;
+		[ ContextualBalloon.pluginName ]: ContextualBalloon;
+		[ Notification.pluginName ]: Notification;
 	}
 }
