+++ packages/ckeditor5-ui/src/panel/balloon/contextualballoon.ts
@@ -860,9 +860,3 @@ class FakePanelsView extends View {
 		}
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ ContextualBalloon.pluginName ]: ContextualBalloon;
-	}
-}
